refactor(backend): standardize to entity.type.js naming convention
