+++ backend/controllers/auth.controller.js
@@ -1,6 +1,6 @@
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
-import User from '../models/userModel.js';
+import User from '../models/user.model.js';
 
 export const registerUser = async (req, res) => {
   try {
